IMP: refactoring + diff. dyn vs static
