+++ include/api/cltj_rdf.hpp
@@ -18,7 +18,7 @@ namespace cltj {
 
     template<class Index    = cltj::compact_dyn_ltj,
              class Trait    = ltj::util::trait_size,
-             class Iterator = ltj::ltj_iterator_lite<Index, uint8_t, uint64_t>,
+             class Iterator = ltj::ltj_iterator_dyn<Index, uint8_t, uint64_t>,
              class Veo      = ltj::veo::veo_adaptive<Iterator, Trait>>
     class cltj_rdf {
 
@@ -239,15 +239,15 @@ namespace cltj {
         typedef cltj::cltj_rdf<> cltj_rdf_dyn;
         //Full Tries + Dynamic + VEO global
         typedef cltj::cltj_rdf<cltj::compact_dyn_ltj, ltj::util::trait_size,
-                ltj::ltj_iterator_lite<cltj::compact_dyn_ltj, uint8_t, uint64_t>,
-                ltj::veo::veo_simple<ltj::ltj_iterator_lite<cltj::compact_dyn_ltj, uint8_t, uint64_t>>> cltj_rdf_dyn_global;
+                ltj::ltj_iterator_dyn<cltj::compact_dyn_ltj, uint8_t, uint64_t>,
+                ltj::veo::veo_simple<ltj::ltj_iterator_dyn<cltj::compact_dyn_ltj, uint8_t, uint64_t>>> cltj_rdf_dyn_global;
         //Meta Tries + Dynamic + VEO adaptive
         typedef cltj::cltj_rdf<cltj::compact_ltj_metatrie_dyn, ltj::util::trait_size,
-                ltj::ltj_iterator_metatrie<cltj::compact_ltj_metatrie_dyn, uint8_t, uint64_t>> xcltj_rdf_dyn;
+                ltj::ltj_iterator_metatrie_dyn<cltj::compact_ltj_metatrie_dyn, uint8_t, uint64_t>> xcltj_rdf_dyn;
         //Meta Tries + Dynamic + VEO global
         typedef cltj::cltj_rdf<cltj::compact_ltj_metatrie_dyn, ltj::util::trait_size,
-                ltj::ltj_iterator_metatrie<cltj::compact_ltj_metatrie_dyn, uint8_t, uint64_t>,
-                ltj::veo::veo_simple<ltj::ltj_iterator_metatrie<cltj::compact_ltj_metatrie_dyn, uint8_t, uint64_t>>> xcltj_rdf_dyn_global;
+                ltj::ltj_iterator_metatrie_dyn<cltj::compact_ltj_metatrie_dyn, uint8_t, uint64_t>,
+                ltj::veo::veo_simple<ltj::ltj_iterator_metatrie_dyn<cltj::compact_ltj_metatrie_dyn, uint8_t, uint64_t>>> xcltj_rdf_dyn_global;
 
 }
 #endif //CLTJ_RDF_HPP
