+++ include/index/cltj_index_spo_dyn.hpp
@@ -65,8 +65,9 @@ namespace cltj {
                 std::sort(beg, end, comparator_order(i));
                 std::vector<uint32_t> syms;
                 std::vector<size_type> lengths;
-                helper::sym_level(beg, end, spo_orders[i], i % 2, syms, lengths);
-                if(i % 2 == 0){
+                auto beg_level = (i & 0x1);
+                helper::sym_level(beg, end, spo_orders[i], beg_level, 3, syms, lengths);
+                if(i & 0x1 == 0){
                     m_gaps[i/2] = lengths[0];
                 }
                 m_tries[i] = trie_type(syms, lengths);
@@ -117,7 +118,7 @@ namespace cltj {
         bool insert(const spo_triple &triple) {
             if(!m_n_triples) {
                 for(size_type i = 0; i < m_tries.size(); ++i) {
-                    m_tries[i] = trie_type(triple, spo_orders[i], i & 0x1);
+                    m_tries[i] = trie_type(triple, spo_orders[i], i & 0x1, 3);
                 }
                 m_gaps = {1,1,1};
                 ++m_n_triples;
