+++ src/bench/bench-query-cltj-rdf.cpp
@@ -58,7 +58,7 @@ void query(const std::string &file, const std::string &queries, const uint64_t l
     uint64_t nQ = 0;
 
     //::util::time::usage::usage_type start, stop;
-    typedef ltj::ltj_iterator_lite<index_scheme_type, uint8_t, uint64_t> iterator_type;
+    typedef ltj::ltj_iterator_dyn<index_scheme_type, uint8_t, uint64_t> iterator_type;
 #if ADAPTIVE
     typedef ltj::ltj_algorithm<iterator_type,
         ltj::veo::veo_adaptive<iterator_type, trait_type> > algorithm_type;
