+++ src/test/interactive-metatrie.cpp
@@ -11,7 +11,7 @@
 template<class index_scheme_type, class trait_type>
 void query(std::vector<ltj::triple_pattern> &q, index_scheme_type &graph) {
 
-        typedef ltj::ltj_iterator_metatrie<index_scheme_type, uint8_t, uint64_t> iterator_type;
+        typedef ltj::ltj_iterator_metatrie_dyn<index_scheme_type, uint8_t, uint64_t> iterator_type;
 #if ADAPTIVE
         typedef ltj::ltj_algorithm<iterator_type,
             ltj::veo::veo_adaptive<iterator_type, trait_type> > algorithm_type;
