+++ src/test/test-mt-dyn.cpp
@@ -25,7 +25,7 @@ template<class index_scheme_type, class trait_type>
 void query_upq(const std::string &index, const std::vector<query_type> &queries, const std::vector<update_type> &updates,
            const uint64_t limit, const uint64_t upq) { //upq = updates per query
 
-    typedef ltj::ltj_iterator_metatrie<index_scheme_type, uint8_t, uint64_t> iterator_type;
+    typedef ltj::ltj_iterator_metatrie_dyn<index_scheme_type, uint8_t, uint64_t> iterator_type;
 #if ADAPTIVE
     typedef ltj::ltj_algorithm<iterator_type,
         ltj::veo::veo_adaptive<iterator_type, trait_type> > algorithm_type;
@@ -74,7 +74,7 @@ void query_upq(const std::string &index, const std::vector<query_type> &queries,
 template<class index_scheme_type, class trait_type>
 void test(const std::string &index, const std::vector<update_type> &updates) { //qpu = queries per update
 
-    typedef ltj::ltj_iterator_metatrie<index_scheme_type, uint8_t, uint64_t> iterator_type;
+    typedef ltj::ltj_iterator_metatrie_dyn<index_scheme_type, uint8_t, uint64_t> iterator_type;
 #if ADAPTIVE
     typedef ltj::ltj_algorithm<iterator_type,
         ltj::veo::veo_adaptive<iterator_type, trait_type> > algorithm_type;
